Add service that allows to trigger an event replay
Introduce event handler class SprintEventLoggingListener which is
assigned to a tracking event processor in order to be used for
event replays. An event replay can be triggered by calling method
EventReplayService#replay(processorName)

For some reason (maybe a possible bug), after performing an event
replay, the handler may be invoked twice when a single event is
fetched by the tracking processor thread (see
https://groups.google.com/forum/#!topic/axonframework/-ncu0jIH-vM)
Therefore (as a temorary solution), a message handler interceptor
is registered with the processor, detecting and dismissing such a
double tracked event.